updated version

+++ TodoApp/routers/todos.py
@@ -4,7 +4,7 @@ from typing import Annotated
 from sqlalchemy.orm import Session
 from models import Todos
 from pydantic import BaseModel, Field
-
+from .auth import get_current_user
 
 router = APIRouter()
 
@@ -19,6 +19,7 @@ def get_db():
 
 #This line is used in FastAPI to inject a database session into route functions
 db_D = Annotated[Session, Depends(get_db)]
+user_D = Annotated[dict, Depends(get_current_user)]
 
 class TodoRequest(BaseModel):
     title: str = Field(min_length=3)
@@ -28,27 +29,40 @@ class TodoRequest(BaseModel):
 
 
 @router.get("/", status_code=status.HTTP_200_OK)
-async def read_all(db: db_D):
-    return db.query(Todos).all()
+async def read_all(user:user_D, db: db_D):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication failed')
+    return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
 
 @router.get("/todo/{todo_id}", status_code = status.HTTP_200_OK)
-async def read_todo(db: db_D, todo_id:int = Path(gt=0)):
+async def read_todo(user:user_D, db: db_D, todo_id:int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication failed')
     #the first() is filter saves the performance time
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+        .filter(Todos.owner_id == user.get('id')).first()
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=404, detail='The todos id is not present')
 
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def post_to_do(db: db_D, todo_request:TodoRequest):
-    todo_model = Todos(**todo_request.dict())
+async def post_to_do(user:user_D, db: db_D,
+                     todo_request:TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication failed')
+    todo_model = Todos(**todo_request.dict(), owner_id = user.get('id'))
 
     db.add(todo_model)
     db.commit()
 
 @router.put('/todo/{todo_id}', status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db:db_D, todo_request :TodoRequest, todo_id :int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def update_todo(user:user_D, db:db_D, 
+                      todo_request :TodoRequest, 
+                      todo_id :int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication failed')
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+        .filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found')
 
@@ -61,11 +75,14 @@ async def update_todo(db:db_D, todo_request :TodoRequest, todo_id :int = Path(gt
     db.commit()
 
 @router.delete('/todo/{todo_id}')
-async def delete_todo(db:db_D,todo_request:TodoRequest,todo_id :int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user:user_D, db:db_D,todo_request:TodoRequest,todo_id :int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+        .filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todos not found')
-    db.query(Todos).filter(Todos.id == todo_id).delete() 
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete() 
     db.commit()
